+++ src/support/mod.rs
@@ -1,6 +1,8 @@
 mod ahash;
 mod arcstr;
+mod chrono;
 mod fxhash;
 mod hashbrown;
 mod ordered_float;
 mod rust_decimal;
+mod time;